requirements.txt
# Baseline dependencies for future parsing/reporting tasks
numpy>=1.26
pandas>=2.2
pydantic>=2.7
pyyaml>=6.0
jinja2>=3.1
bibtexparser>=1.4
matplotlib>=3.8
python-pptx>=1.0.2
Pillow>=10.0
